+++ hpx/runtime/threads/executors/current_executor.hpp
@@ -10,7 +10,7 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_THREAD_EXECUTORS_COMPATIBILITY)
-#include <hpx/parallel/executors/thread_pool_executor.hpp>
+#include <hpx/execution/executors/thread_pool_executor.hpp>
 
 namespace hpx { namespace threads { namespace executors {
     //  TODO: Keep get_state?
