+++ libs/execution/include/hpx/execution/executors/current_executor.hpp
@@ -7,7 +7,7 @@
 #ifndef HPX_PARALLEL_EXECUTORS_CURRENT_EXECUTOR_HPP
 #define HPX_PARALLEL_EXECUTORS_CURRENT_EXECUTOR_HPP
 
-#include <hpx/parallel/executors/thread_pool_executor.hpp>
+#include <hpx/execution/executors/thread_pool_executor.hpp>
 
 namespace hpx { namespace parallel { namespace execution {
     using current_executor = parallel::execution::thread_pool_executor;
