+++ libs/execution/include/hpx/execution/executors/guided_pool_executor.hpp
@@ -13,7 +13,7 @@
 #include <hpx/functional/bind_back.hpp>
 #include <hpx/functional/invoke.hpp>
 #include <hpx/lcos/dataflow.hpp>
-#include <hpx/parallel/executors/pool_executor.hpp>
+#include <hpx/execution/executors/pool_executor.hpp>
 #include <hpx/traits/is_future_tuple.hpp>
 #include <hpx/util/pack_traversal.hpp>
 #include <hpx/util/thread_description.hpp>
