+++ libs/execution/include/hpx/execution/executors/limiting_executor.hpp
@@ -10,7 +10,7 @@
 #include <hpx/config.hpp>
 #include <hpx/apply.hpp>
 #include <hpx/functional/invoke.hpp>
-#include <hpx/parallel/executors/execution_fwd.hpp>
+#include <hpx/execution/executors/execution_fwd.hpp>
 #include <hpx/runtime/threads/executors/default_executor.hpp>
 #include <hpx/runtime/threads/executors/pool_executor.hpp>
 #include <hpx/runtime/threads/thread_executor.hpp>
