+++ tests/regressions/lcos/shared_future_then_with_executor_3634.cpp
@@ -12,7 +12,7 @@
 #include <hpx/lcos/future.hpp>
 
 #include <hpx/execution/executors.hpp> // Workaround for a missing header file
-#include <hpx/parallel/executors/pool_executor.hpp>
+#include <hpx/execution/executors/pool_executor.hpp>
 
 int main()
 {
